+++ angular/src/app/components/moms/moms.component.ts
@@ -1,3 +1,4 @@
+import { ConfigStateService } from '@abp/ng.core';
 import { NgFor } from '@angular/common';
 import { Component, Input, OnInit } from '@angular/core';
 import {
@@ -38,9 +39,10 @@ export class MomsComponent implements OnInit {
   constructor(
     private fb: FormBuilder,
     private route: ActivatedRoute,
-    private meetingMinuteService: MomService
+    private meetingMinuteService: MomService,
+    private config: ConfigStateService
   ) {
-    this.projectId = this.route.snapshot.pathFromRoot[1].params['id'];
+    this.projectId = this.route.snapshot.pathFromRoot[2].params['id'];
   }
 
   ngOnInit(): void {
@@ -53,14 +55,18 @@ export class MomsComponent implements OnInit {
       error => {
         this.addExistingData([]);
         if (error.status == 403) {
-          this.unauthorizedPerson = false;
-          console.log(this.unauthorizedPerson);
           console.warn('Unauthorized access (403):', error);
         } else {
           console.error('Error fetching projects:', error);
         }
       }
     );
+    if (
+      this.config.getOne('currentUser').roles[0] == 'client' ||
+      this.config.getOne('currentUser').roles[0] == 'auditor'
+    ) {
+      this.unauthorizedPerson = true;
+    }
   }
 
   addExistingData(data: any): void {
@@ -105,6 +111,10 @@ export class MomsComponent implements OnInit {
   }
 
   removeRow(index: number): void {
+    const getConfirmation = window.confirm('Do you want to delte');
+    if (getConfirmation == false) {
+      return;
+    }
     const approveteamArray = this.forms.get('formitem') as FormArray;
     const controlAtIndex = approveteamArray.at(index);
     this.meetingMinuteService.deleteItem(controlAtIndex.value.id).subscribe(
@@ -119,7 +129,7 @@ export class MomsComponent implements OnInit {
   }
 
   onSubmit(): void {
-    if (this.forms.valid) {
+    if (this.forms.valid && this.unauthorizedPerson==false) {
       this.forms.value.formitem.forEach(async e => {
         try {
           const modelDate: MeetingMinute = {
@@ -141,9 +151,8 @@ export class MomsComponent implements OnInit {
           console.error('Error:', error);
         }
       });
-    }
-    else{
-      alert("Make sure you filled right value, check date formate");
+    } else {
+      alert('Make sure you filled right value, check date formate');
     }
   }
 }
